backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
easyocr
pymupdf
pillow
numpy
scikit-learn
pandas
joblib
spacy
torch
openai
